refactor: Redux の slice 追加の前準備として、features ディレクトリの構造を変更してサポカ関連を features/Support に集約した
