+++ src/features/Support/index.tsx
@@ -1,0 +1,18 @@
+import { VStack } from '@chakra-ui/react';
+import React from 'react';
+
+import CardDeck from './CardDeck';
+import Skills from './Skills';
+import Status from './Status';
+
+const Support = () => {
+  return (
+    <VStack w='100%' gap='4'>
+      <CardDeck />
+      <Skills />
+      <Status />
+    </VStack>
+  );
+};
+
+export default Support;
