+++ src/pages/index.tsx
@@ -1,10 +1,8 @@
-import { Container, Text, VStack } from '@chakra-ui/react';
+import { Container, VStack } from '@chakra-ui/react';
 import type { NextPageWithLayout } from 'next';
 import Head from 'next/head';
 
-import CardDeck from '@/features/CardDeck';
-import Skills from '@/features/Skills';
-import Status from '@/features/Status';
+import Support from '@/features/Support';
 import Test from '@/features/Test';
 import { Layout } from '@/layouts/Layout';
 
@@ -18,10 +16,7 @@ const Home: NextPageWithLayout = () => (
 
     <Container as='main' maxW='container.lg' bgColor='gray.100'>
       <VStack p='4'>
-        <CardDeck />
-        <Status />
-        <Skills />
-        <Text>hello</Text>
+        <Support />
         <Test />
       </VStack>
     </Container>
